Add "game" components to display minesweeper (like) game board
The "game" is based on the API provided by
https://github.com/bryanjswift/the-mines-challenge. It defines endpoints
for listing games, viewing a specific game's state and executing a
"move" in the game. The game listing screen usings the API during server
side rendering to fetch the list of game ids and display them.

Moves `route_dispatch` module functions into a `RouteDispatcher` struct.
The `RouteDispatcher` maintains a reference counted `Transmitter` to
pass into newly constructed `GameList` components. The `Transmitter` is
needed to enable navigation to a new `Route` from the `GameList`.

Changes to the game are sent through the `api::patch_game` function
which sends a PATCH to the server, the server responds with an updated
board state. The board view is replaced on each update.